separate UserController from ApplicationController
