+++ app/views/registration.scala.html
@@ -53,7 +53,7 @@
 				<div class="col-sm-4 col-sm-offset-1">
 					<div class="signup-form"><!--sign up form-->
 						<h2>@message</h2>
-						<form action="@routes.Application.register" method ="POST">
+						<form action="@routes.UserController.register" method ="POST">
 							<input type="text" id="username" name="username" placeholder="Name"/>	
 							<input type="email" id="email" name="email" placeholder="Email Address"/>
 							<input type="password" id="password" name="password" placeholder="Password"/>
