+++ bin/evolutions/default/1.sql
@@ -4,7 +4,7 @@
 # --- !Ups
 
 create table user (
-  id                        integer not null,
+  id                        bigint not null,
   username                  varchar(255),
   email                     varchar(255),
   password                  varchar(255),
