+++ bin/views/Loginpage.scala.html
@@ -51,7 +51,7 @@
 				<div class="col-sm-4 col-sm-offset-1">
 					<div class="login-form"><!--login form-->
 						<h2>@info</h2>
-						<form action="@routes.Application.login" method ="POST">
+						<form action="@routes.UserController.login" method ="POST">
 							<input type="text" id="email" name="email" placeholder="Email" />
 							<input type="password" id="password" name="password" placeholder="Password" />
 							<button type="submit" class="btn btn-primary">Login</button>
